Added ModuleDetailValidatorMixin
