+++ corehq/apps/app_manager/models.py
@@ -155,7 +155,6 @@ from corehq.apps.app_manager.exceptions import (
     XFormIdNotUnique,
     XFormValidationError,
     ScheduleError,
-    CaseXPathValidationError,
     XFormValidationFailed,
     PracticeUserException)
 from corehq.apps.reports.daterange import get_daterange_start_end_dates, get_simple_dateranges
@@ -2764,46 +2763,6 @@ class ModuleDetailsMixin(object):
         if module_offers_search(self) and not self.case_details.short.custom_xml:
             details.append(('search_short', self.search_detail, True))
         return tuple(details)
-
-    def validate_details_for_build(self):
-        errors = []
-        for sort_element in self.detail_sort_elements:
-            try:
-                validate_detail_screen_field(sort_element.field)
-            except ValueError:
-                errors.append({
-                    'type': 'invalid sort field',
-                    'field': sort_element.field,
-                    'module': self.get_module_info(),
-                })
-        if self.case_list_filter:
-            try:
-                # test filter is valid, while allowing for advanced user hacks like "foo = 1][bar = 2"
-                case_list_filter = interpolate_xpath('dummy[' + self.case_list_filter + ']')
-                etree.XPath(case_list_filter)
-            except (etree.XPathSyntaxError, CaseXPathValidationError):
-                errors.append({
-                    'type': 'invalid filter xpath',
-                    'module': self.get_module_info(),
-                    'filter': self.case_list_filter,
-                })
-        for detail in [self.case_details.short, self.case_details.long]:
-            if detail.use_case_tiles:
-                if not detail.display == "short":
-                    errors.append({
-                        'type': "invalid tile configuration",
-                        'module': self.get_module_info(),
-                        'reason': _('Case tiles may only be used for the case list (not the case details).')
-                    })
-                col_by_tile_field = {c.case_tile_field: c for c in detail.columns}
-                for field in ["header", "top_left", "sex", "bottom_left", "date"]:
-                    if field not in col_by_tile_field:
-                        errors.append({
-                            'type': "invalid tile configuration",
-                            'module': self.get_module_info(),
-                            'reason': _('A case property must be assigned to the "{}" tile field.'.format(field))
-                        })
-        return errors
 
     def get_case_errors(self, needs_case_type, needs_case_detail, needs_referral_detail=False):
         module_info = self.get_module_info()
